bam/tui/pyproject.toml
[project]
name = "bam"
version = "0.1.0"
description = "BAM - Bioimage Analysis Manager TUI"
requires-python = ">=3.11"
dependencies = [
    "textual>=0.40",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.scripts]
bam = "ba_tui.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ba_tui"]
